Save Commit

+++ ft_printf/rendu_ft_printf/ft_strlen.c
@@ -12,7 +12,7 @@
 
 #include "libftprintf.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	size_t	i;
